implement complex objects
implemented code for complex object return types

Struggled with the RunTimeException test since i did throw it when something went wrong but the problem was that nothing went wrong.

i analyzed the test and saw that there was a callmessage param set but the method doesn't have any parameters. I solved this by adding a counter that keeps track of the amount of params it gets and when that is different from the initial params, it throws an error which will result in the RunTimeException.